add cart, register

+++ src/main/java/controlllers/user/GioHangUserServlet.java
@@ -16,42 +16,25 @@ import java.sql.Date;
 import java.time.LocalDate;
 import java.util.List;
 
-@WebServlet("/GioHangUserServlet/*")
+@WebServlet({"/GioHangUserServlet/index", "/GioHangUserServlet/store", "/GioHangUserServlet/update", "/GioHangUserServlet/delete"})
 public class GioHangUserServlet extends HttpServlet {
     private GioHangRepository gioHangRepository = new GioHangRepository();
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String path = request.getPathInfo();
-        if (path == null) {
-            path = "/index";
-        }
-        switch (path) {
-            case "/index":
-            case "/create":
-            case "/edit":
-                showGioHang(request, response);
-                break;
-            case "/delete":
-                deleteGioHang(request, response);
-                break;
-            default:
-                showGioHang(request, response);
-                break;
+        String uri = request.getRequestURI();
+        if (uri.contains("index")) {
+            showGioHang(request, response);
+        } else if (uri.contains("delete")) {
+            deleteGioHang(request, response);
         }
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String path = request.getPathInfo();
-        switch (path) {
-            case "/store":
-                storeGioHang(request, response);
-                break;
-            case "/update":
-                updateGioHang(request, response);
-                break;
-            default:
-                showGioHang(request, response);
-                break;
+        String uri = request.getRequestURI();
+        if (uri.contains("store")) {
+            storeGioHang(request, response);
+        } else if (uri.contains("update")) {
+            updateGioHang(request, response);
         }
     }
 
@@ -59,16 +42,6 @@ public class GioHangUserServlet extends HttpServlet {
         HttpSession session = request.getSession();
         KhachHang khachHang = (KhachHang) session.getAttribute("user");
         GioHang gioHang = gioHangRepository.getGioHang(khachHang.getId());
-        if (gioHang == null) {
-            gioHang = new GioHang();
-            gioHang.setKhachHang(khachHang);
-            int year = LocalDate.now().getYear();
-            int month = LocalDate.now().getMonthValue();
-            int day = LocalDate.now().getDayOfMonth();
-            gioHang.setNgayTao(Date.valueOf(LocalDate.of(year, month, day)));
-            gioHang.setTrangThai(0);
-            gioHangRepository.insertGioHang(gioHang);
-        }
         List<GioHangChiTietViewModel> listGioHangChiTiet = gioHangRepository.getAllGioHangChiTietByGioHangId(gioHang.getId());
 
         String realPath = request.getServletContext().getRealPath("/images");
@@ -80,7 +53,6 @@ public class GioHangUserServlet extends HttpServlet {
                 gioHangChiTietViewModel.setSrcImage(request.getContextPath() + "/images/" + fileName);
             }
         }
-        request.setAttribute("gioHang", gioHang);
         request.setAttribute("listGioHangChiTiet", listGioHangChiTiet);
         request.setAttribute("view", "/views/user/gioHang/gioHang.jsp");
         request.getRequestDispatcher("/views/user/layoutUser.jsp").forward(request, response);
@@ -91,21 +63,12 @@ public class GioHangUserServlet extends HttpServlet {
         HttpSession session = request.getSession();
         KhachHang khachHang = (KhachHang) session.getAttribute("user");
         GioHang gioHang = gioHangRepository.getGioHang(khachHang.getId());
-        if (gioHang == null) {
-            gioHang = new GioHang();
-            gioHang.setKhachHang(khachHang);
-            int year = LocalDate.now().getYear();
-            int month = LocalDate.now().getMonthValue();
-            int day = LocalDate.now().getDayOfMonth();
-            gioHang.setNgayTao(Date.valueOf(LocalDate.of(year, month, day)));
-            gioHang.setTrangThai(0);
-            gioHangRepository.insertGioHang(gioHang);
-        }
         String idChiTietSp = request.getParameter("id");
+        request.getParameter("soLuong");
         int soLuong = Integer.parseInt(request.getParameter("soLuong"));
-        GioHangChiTiet gioHangChiTiet = gioHangRepository.getGioHangChiTietByIdGioHang(gioHang.getId(), idChiTietSp);
-        if (gioHangChiTiet == null) {
-            gioHangChiTiet = new GioHangChiTiet();
+        List<GioHangChiTiet> list = gioHangRepository.getGioHangChiTietByIdGioHang(gioHang.getId(), idChiTietSp);
+        if (list.isEmpty()) {
+            GioHangChiTiet gioHangChiTiet = new GioHangChiTiet();
             gioHangChiTiet.setGioHang(gioHang);
             ChiTietSp chiTietSp = gioHangRepository.getChiTietSp(idChiTietSp);
             gioHangChiTiet.setChiTietSp(chiTietSp);
@@ -113,22 +76,13 @@ public class GioHangUserServlet extends HttpServlet {
             gioHangChiTiet.setSoLuongTon(soLuong);
             gioHangRepository.insertGioHangChiTiet(gioHangChiTiet);
         } else {
+            GioHangChiTiet gioHangChiTiet = list.get(0);
             gioHangChiTiet.setSoLuongTon(gioHangChiTiet.getSoLuongTon() + soLuong);
             gioHangRepository.updateGioHangChiTiet(gioHangChiTiet);
         }
         response.sendRedirect(request.getContextPath() + "/GioHangUserServlet/index");
     }
 
-    protected void deleteGioHang(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        HttpSession session = request.getSession();
-        KhachHang khachHang = (KhachHang) session.getAttribute("user");
-        String idGioHangChiTiet = request.getParameter("id");
-        GioHangChiTiet gioHangChiTiet = gioHangRepository.getGioHangChiTietById(idGioHangChiTiet);
-        if (gioHangChiTiet != null) {
-            gioHangRepository.deltete(gioHangChiTiet);
-        }
-        response.sendRedirect(request.getContextPath() + "/GioHangUserServlet/index");
-    }
 
     protected void updateGioHang(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         HttpSession session = request.getSession();
@@ -140,7 +94,17 @@ public class GioHangUserServlet extends HttpServlet {
             gioHangChiTiet.setSoLuongTon(soLuong);
             gioHangRepository.updateGioHangChiTiet(gioHangChiTiet);
         }
-        response.sendRedirect(request.getContextPath() + "/GioHangUserServlet/index");
+        response.sendRedirect("/Assignment_Sof3011_war_exploded/GioHangUserServlet/index");
+    }
+    protected void deleteGioHang(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        HttpSession session = request.getSession();
+        KhachHang khachHang = (KhachHang) session.getAttribute("user");
+        String id = request.getParameter("id");
+        GioHangChiTiet gioHangChiTiet = gioHangRepository.getGioHangChiTietById(id);
+        GioHang gioHang = gioHangRepository.getGioHang(khachHang.getId());
+            gioHangRepository.delete(id, gioHang.getId());
+            response.sendRedirect("/Assignment_Sof3011_war_exploded/GioHangUserServlet/index");
+
     }
 
 }
